refactor(iocuak-core): move loadContainerModule metadata actions to containerModuleMetadata
