+++ packages/iocuak-core/src/containerModuleMetadata/actions/loadContainerModuleElementAsync.ts
@@ -1,5 +1,5 @@
 import { ContainerModuleMetadata } from '../../containerModuleMetadata/models/ContainerModuleMetadata';
-import { CreateInstanceTaskContext } from '../models/CreateInstanceTaskContext';
+import { CreateInstanceTaskContext } from '../../createInstanceTask/models/CreateInstanceTaskContext';
 import { loadContainerModuleElement } from './loadContainerModuleElement';
 
 export async function loadContainerModuleElementAsync(
